Fixed values of N so program runs smoothly with doubles

+++ HW1/src/Sum.java
@@ -1,62 +1,69 @@
 
+// Implements six loops that return sums from Exercise 2_7 in Weiss textbook
 public class Sum {
 	
-	public int sumOne(int n) {
-		int sum = 0;
-		for ( int i = 0; i < n; i++ ) {
+	// Expected O(N)
+	public double sumOne(double n) {
+		double sum = 0;
+		for ( double i = 0; i < n; i++ ) {
 			sum++;
 		}
 		return sum;
 	}
 	
-	public int sumTwo(int n) {
-		int sum = 0;
-		for ( int i = 0; i < n; i++ ) {
-			for ( int j = 0; j < n; j++ ) {
+	// Expected O(N^2)
+	public double sumTwo(double n) {
+		double sum = 0;
+		for ( double i = 0; i < n; i++ ) {
+			for ( double j = 0; j < n; j++ ) {
 				sum++;
 			}
 		}
 		return sum;
 	}
 	
-	public int sumThree(int n) {
-		int sum = 0;
-        for ( int i = 0; i < n; i++ ) {
-            for ( int j = 0; j < n * n; j++ ) {
+	// Expected O(N^3)
+	public double sumThree(double n) {
+		double sum = 0;
+        for ( double i = 0; i < n; i++ ) {
+            for ( double j = 0; j < n * n; j++ ) {
                 sum++;
             }
         }
         return sum;
 	}
 
-	public int sumFour(int n) {
-		int sum = 0;
-        for ( int i = 0; i < n; i++ ) {
-            for ( int j = 0; j < i; j++ ) {
+	// Expected O(N^2)
+	public double sumFour(double n) {
+		double sum = 0;
+        for ( double i = 0; i < n; i++ ) {
+            for ( double j = 0; j < i; j++ ) {
                 sum++;
             }
         }
         return sum;
 	}
 	
-	public int sumFive(int n) {
-		int sum = 0;
-		for ( int i = 0; i < n; i++ ) {
-			for ( int j = 0; j < (i * i); j++ ) {
-		    	 for (int k = 0; k < j; k++ ) {
+	// Expected O(N^5)
+	public double sumFive(double n) {
+		double sum = 0;
+		for ( double i = 0; i < n; i++ ) {
+			for ( double j = 0; j < (i * i); j++ ) {
+		    	 for ( double k = 0; k < j; k++ ) {
 		                    	   		sum++;
 		    	 }
 			}
 		}
 		return sum;
 	}
-		
-	public int sumSix(int n) {
-		int sum = 0;
-		for ( int i = 1; i < n; i++ ) {
-			for ( int j = 1; j < i * i; j++ ) {
+	
+	// Expected O(N^4)
+	public double sumSix(double n) {
+		double sum = 0;
+		for ( double i = 1; i < n; i++ ) {
+			for ( double j = 1; j < i * i; j++ ) {
 				if ( j % i == 0 ) {
-					for ( int k = 0; k < j; k++ ) {
+					for ( double k = 0; k < j; k++ ) {
 		                              sum++;
 					}
 				}
@@ -65,6 +72,4 @@ public class Sum {
 		return sum;
 	}
 	
-
-
 }
